+++ src/main/java/frc/robot/RobotContainer.java
@@ -12,6 +12,7 @@ import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import frc.robot.subsystems.Climber;
 import frc.robot.subsystems.DriveTrain;
 import frc.robot.subsystems.Intake;
+import frc.robot.subsystems.IntakePneumatics;
 import frc.robot.subsystems.Shooter;
 import frc.robot.Constants.OIConstants;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
@@ -33,6 +34,7 @@ public class RobotContainer {
   public final DriveTrain m_drivetrain = new DriveTrain();
   public final Climber m_Climber = new Climber();
   public final Intake m_intake = new Intake();
+  public final IntakePneumatics mPneumatics = new IntakePneumatics();
   public final Shooter m_shooter = new Shooter();
   
   public final static XboxController m_driverController = new XboxController(OIConstants.kDriverControllerPort);
@@ -81,10 +83,10 @@ public class RobotContainer {
     .whenReleased(() -> m_intake.stopIntake());
 
     new JoystickButton(m_driver2Controller, Button.kX.value)
-    .whenPressed(() -> m_intake.extendIntake());
+    .whenPressed(() -> mPneumatics.extendIntake());
 
     new JoystickButton(m_driver2Controller, Button.kY.value)
-    .whenPressed(() -> m_intake.retractIntake());
+    .whenPressed(() -> mPneumatics.retractIntake());
 
     //Climber
